Move factories from registrations to events
