+++ registrations/tests/factories.py
@@ -1,33 +1,7 @@
 import factory
 
-from events.models import DataSource, Event
+from events.tests.factories import EventFactory
 from registrations.models import Registration
-
-
-class DataSourceFactory(factory.django.DjangoModelFactory):
-    id = factory.Sequence(lambda n: "data-source-{0}".format(n))
-
-    class Meta:
-        model = DataSource
-
-
-class OrganizationFactory(factory.django.DjangoModelFactory):
-    data_source = factory.SubFactory(DataSourceFactory)
-
-    class Meta:
-        model = "django_orghierarchy.Organization"
-
-
-class EventFactory(factory.django.DjangoModelFactory):
-    data_source = factory.SubFactory(DataSourceFactory)
-    publisher = factory.SubFactory(OrganizationFactory)
-
-    class Meta:
-        model = Event
-
-    @factory.lazy_attribute_sequence
-    def id(self, n):
-        return f"{self.data_source.id}:{n}"
 
 
 class RegistrationFactory(factory.django.DjangoModelFactory):
